Change the way an EnOcean Switch is updated
Now a Switch sensor is not updated en the Switch event binded to Ubikit switch event system
but in the log method in the ubikitAdapter.

With this modification the meaning of update state message change.
The switchNumber code the button(s) that has been pressed and the
button status code the state of the button (true=pressed, false=released).