#144 HMI light sensor : Display in device + server
Server : Now enum to represent the scale of the sensor, it has
functions to get the different information of the scale, add label in
JSON of getDescription + notifychange("label") in same time of the
value.
HMI device, label in list device + deviceDetails (label : value +
scale), change needed in the menu.
Scale : If want to modify, change in server and in illumination.html.