Fix issue #147 : Unlocated device not alphabetic sort
In Device, menu and list, the unlocated place will not be in its
alphabetic place but at the end of the list of places.
No change at the place sorting but just modify the html to show in
first the places except the unlocated place. and after the unlocated
place.